Seperated handler logic
moved all handler logic from main to appropriate files
